% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictogram_extractor.R
\name{verify_timestamps}
\alias{verify_timestamps}
\title{Verify Timestamps of Extracted Frames}
\usage{
verify_timestamps(extractor, ...)
}
\arguments{
\item{extractor}{An object of class \code{VideoFrameExtractor}.}

\item{...}{Additional arguments.}
}
\value{
A \code{data.frame} (returned invisibly) containing metadata extracted
from the frames, such as original and modified timestamps. If no frames
have been extracted, returns \code{NULL} invisibly with a message.
}
\description{
Reads back the metadata from the extracted frames to verify it was correctly stamped.
}
