% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictogram_extractor.R
\name{extract}
\alias{extract}
\title{Extract Frames from Video or Folder}
\usage{
extract(extractor, verbose = TRUE, ...)
}
\arguments{
\item{extractor}{An object of class \code{VideoFrameExtractor} or \code{FolderExtractor}.}

\item{verbose}{logical. If TRUE, progress messages are printed.}

\item{...}{Additional arguments.}
}
\value{
The original extractor object (either \code{VideoFrameExtractor} or
\code{FolderExtractor}) returned invisibly. The object is updated with the
paths to the extracted frames or the results of the batch processing.
}
\description{
Generic function to extract frames from a video file or all videos in a folder.
}
